% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{circularLegends}
\alias{circularLegends}
\title{Circular legend markers for ggplot line or bar charts}
\usage{
circularLegends(g = g)
}
\arguments{
\item{g}{A ggplot object. Defaults to g}
}
\description{
This function removes standard item legends and adds circular legend items mapper to aes(color). Note that if aes(fill) is also mapped, 'guide=FALSE' must be added to to scale_fill_XX.
}

