% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{smartRound}
\alias{smartRound}
\title{A smarter rounding function}
\usage{
smartRound(x, digits = 1)
}
\arguments{
\item{digits}{Number of decimal places to round to. Defaults to 1.}
}
\description{
This function rounds without truncating trailing zeros.
}

