% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{theme_mhc}
\alias{theme_mhc}
\title{A ggplot2 theme for the Commission}
\usage{
theme_mhc(base_size = 11, base_family = "VAG", ticks = TRUE,
  truncate = FALSE, xTitle = TRUE, yTitle = TRUE)
}
\arguments{
\item{ticks}{Include axis ticks. Defaults to TRUE.}

\item{truncate}{Treat y axis as truncated (hide x axis line). Defaults to FALSE.}

\item{xTitle}{Include x axis title. Defaults to TRUE}

\item{yTitle}{Include y axis title. Defaults to TRUE}
}
\description{
This function ensures all plots conform to the same standards
}

